requirements.txt
flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.10
waitress==2.1.2
gunicorn==21.2.0

# 通知系统依赖
requests==2.32.4
